+++ extensions/healthie/actions/index.ts
@@ -18,6 +18,8 @@ import { updatePatient } from './updatePatient'
 import { getPatient } from './getPatient'
 import { applyTagToPatient } from './applyTagToPatient'
 import { getMetricEntry } from './getMetricEntry'
+import { updatePatientQuickNote } from './updatePatientQuickNote'
+import { createMetricEntry } from './createMetricEntry'
 
 export const actions = {
   createAppointment,
@@ -49,4 +51,6 @@ export const actions = {
   completeTask,
   assignPatientToGroup,
   getMetricEntry,
+  updatePatientQuickNote,
+  createMetricEntry,
 }