+++ extensions/cloudinary/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Cloudinary changelog
+
+## October 2023
+
+- Action added: Upload a single file